feat: add cache to make some stuff faster
